+++ BancoFicV2/Program.cs
@@ -15,7 +15,7 @@ namespace BancoFicV2
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Opcoesiniciais());
+            Application.Run(new OpcoesIniciais());
         }
     }
 }
